Update dependencies to use php-semver-checker ^0.14
Add symfony/process with constraint ^3.4 in order for the
klaussilveira/gitter dependency to limit symfony/process to a version
that still supports ProcessUtils::escapeArguments. While it would be
possible to update this dependency, it is instead easier to add a
constraint on this dependency.

Set vierbergenlars/php-semver to ^3.0.1 as 3.0.0 does not support PHP
7.2 changes regarding the use of "string" as a class name.